Rephased baselines are joined up when plotted as a single vector. This means that a line will join the end of a base line to the beginning of another. Visually, this isn't very useful. In order to avoid this, we split baselines into dataframes with a column for each rephased value. split is a utility function which splits a vector into a dataframe for this purpose.

Rephased baselines are joined up when plotted as a single vector. This means that a line will join the end of a base line to the beginning of another. Visually, this isn't very useful. In order to avoid this, we split baselines into dataframes with a column for each rephased value. split is a utility function which splits a vector into a dataframe for this purpose.

split(v, vname)

Arguments

v
a vector for split to split into a dataframe.

vname
the columns of the returned dataframe are named vname1, vname2, vname3, etc

Value

A data.frame with a column for each new value of v.

See also

sus multi_shift

Examples

runchart:::split(
v = c(1,1,1,1,1,1,5,5,5,5,5,5,5,5,3,3,3,3,3,3,3),
vname = "example"
)

#>    example1 example2 example3
#> 1         1       NA       NA
#> 2         1       NA       NA
#> 3         1       NA       NA
#> 4         1       NA       NA
#> 5         1       NA       NA
#> 6         1       NA       NA
#> 7        NA        5       NA
#> 8        NA        5       NA
#> 9        NA        5       NA
#> 10       NA        5       NA
#> 11       NA        5       NA
#> 12       NA        5       NA
#> 13       NA        5       NA
#> 14       NA        5       NA
#> 15       NA       NA        3
#> 16       NA       NA        3
#> 17       NA       NA        3
#> 18       NA       NA        3
#> 19       NA       NA        3
#> 20       NA       NA        3
#> 21       NA       NA        3